trigo et invtrigo use cayley_extend
