+++ include/kyosu/types/impl/trigo.hpp
@@ -9,6 +9,7 @@
 
 #include <eve/module/core.hpp>
 #include <eve/module/math.hpp>
+#include <kyosu/details/cayleyify.hpp>
 
 namespace kyosu::_
 {
@@ -34,8 +35,7 @@ namespace kyosu::_
     }
     else
     {
-      auto e = kyosu::exp(z);
-      return kyosu::average(e, rec(e));
+      return cayley_extend(cosh, z);
     }
   }
 
@@ -73,8 +73,7 @@ namespace kyosu::_
     }
     else
     {
-      auto e = kyosu::exp(z);
-      return eve::average(e, -kyosu::rec(e));
+      return cayley_extend(sinh, z);
     }
   }
 
@@ -122,9 +121,7 @@ namespace kyosu::_
     }
     else
     {
-      auto e = kyosu::exp(z);
-      auto ie = kyosu::rec(e);
-      return kumi::tuple{ eve::average(e, -ie),  eve::average(e, ie)};
+      return cayley_extend2(sinhcosh, z);
     }
   }
 
@@ -145,8 +142,7 @@ namespace kyosu::_
     }
     else
     {
-      auto e = kyosu::expm1(z+z);
-      return e/(e+2);
+      return cayley_extend(tanh, z);
     }
   }
 
@@ -160,8 +156,7 @@ namespace kyosu::_
     }
     else
     {
-      auto e = kyosu::expm1(z+z);
-      return (e+2)/e;
+      return cayley_extend(coth, z);
     }
   }
 
@@ -175,11 +170,7 @@ namespace kyosu::_
     }
     else
     {
-      auto p = kyosu::pure(z);
-      auto az = kyosu::abs(p);
-      auto [s, c] = eve::sincos(real(z));
-      auto w = -s*eve::sinhc(az);
-      return c*cosh(az)+w*pure(z);
+      return cayley_extend(cos, z);
     }
   }
 
@@ -201,11 +192,7 @@ namespace kyosu::_
     }
     else
     {
-      auto p = kyosu::pure(z);
-      auto az = kyosu::abs(p);
-      auto [s, c] = eve::sincos(real(z));
-      auto w = c*eve::sinhc(az);
-      return s*cosh(az)+w*pure(z);
+      return cayley_extend(sin, z);
     }
   }
 
@@ -228,16 +215,7 @@ namespace kyosu::_
     }
     else
     {
-      auto p = kyosu::pure(z);
-      auto az = kyosu::abs(p);
-      auto [s, c] = eve::sincos(kyosu::real(z));
-      auto shc = eve::sinhc(az);
-      auto ch  = eve::cosh(az);
-      auto wc = c*shc;
-      auto ws =-s*shc;
-      auto sq = s*ch + wc*p;
-      auto cq = c*ch + ws*p;
-      return kumi::tuple{sq, cq};
+      return cayley_extend2(sincos, z);
     }
   }
 
@@ -252,8 +230,7 @@ namespace kyosu::_
     }
     else
     {
-      auto [s, c] = sincos(z);
-      return s/c;
+      return cayley_extend(tan, z);
     }
   }
 
@@ -268,8 +245,7 @@ namespace kyosu::_
     }
     else
     {
-      auto [s, c] = sincos(z);
-      return c/s;
+      return cayley_extend(cot, z);
     }
   }
 
@@ -277,96 +253,103 @@ namespace kyosu::_
   KYOSU_FORCEINLINE constexpr
   auto dispatch(eve::tag_of<kyosu::sinc> const&, C const& z) noexcept
   {
-    auto s = kyosu::sin(z);
-    using u_t = eve::underlying_type_t<C>;
-    return kyosu::if_else(kyosu::abs(z) < eve::eps(eve::as(u_t())), eve::one(eve::as(u_t())), s/z);
+    if constexpr(concepts::complex<C> )
+    {
+      auto s = kyosu::sin(z);
+      using u_t = eve::underlying_type_t<C>;
+      return kyosu::if_else(kyosu::abs(z) < eve::eps(eve::as(u_t())), eve::one(eve::as(u_t())), s/z);
+    }
+    else
+    {
+      return cayley_extend(sinc, z);
+    }
   }
 
-  template<typename C>
-  KYOSU_FORCEINLINE constexpr
-  auto dispatch(eve::tag_of<kyosu::atanh> const&, C const& z) noexcept
-  {
-    return eve::half(eve::as(real(z)))*(log(inc(z))-log(oneminus(z)));
-  }
+//   template<typename C>
+//   KYOSU_FORCEINLINE constexpr
+//   auto dispatch(eve::tag_of<kyosu::atanh> const&, C const& z) noexcept
+//   {
+//     return eve::half(eve::as(real(z)))*(log(inc(z))-log(oneminus(z)));
+//   }
 
-  template<typename C>
-  KYOSU_FORCEINLINE constexpr
-  auto dispatch(eve::tag_of<kyosu::acoth> const&, C const& z) noexcept
-  {
-    return eve::half(eve::as(real(z)))*(log(inc(z))-log(dec(z)));
-  }
+//   template<typename C>
+//   KYOSU_FORCEINLINE constexpr
+//   auto dispatch(eve::tag_of<kyosu::acoth> const&, C const& z) noexcept
+//   {
+//     return eve::half(eve::as(real(z)))*(log(inc(z))-log(dec(z)));
+//   }
 
-  template<typename C>
-  KYOSU_FORCEINLINE constexpr
-  auto dispatch(eve::tag_of<kyosu::acosh> const&, C const& z) noexcept
-  {
-    return log(z+sqrt(inc(z))*sqrt(dec(z)));
-  }
+//   template<typename C>
+//   KYOSU_FORCEINLINE constexpr
+//   auto dispatch(eve::tag_of<kyosu::acosh> const&, C const& z) noexcept
+//   {
+//     return log(z+sqrt(inc(z))*sqrt(dec(z)));
+//   }
 
-  template<typename C>
-  KYOSU_FORCEINLINE constexpr
-  auto dispatch(eve::tag_of<kyosu::asinh> const&, C const& z) noexcept
-  {
-    return log(z+sqrt(inc(sqr(z))));
-  }
+//   template<typename C>
+//   KYOSU_FORCEINLINE constexpr
+//   auto dispatch(eve::tag_of<kyosu::asinh> const&, C const& z) noexcept
+//   {
+//     return log(z+sqrt(inc(sqr(z))));
+//   }
 
-  template<typename C>
-  KYOSU_FORCEINLINE constexpr
-  auto dispatch(eve::tag_of<kyosu::asech> const&, C const& z) noexcept
-  {
-    return acosh(rec(z));
-  }
+//   template<typename C>
+//   KYOSU_FORCEINLINE constexpr
+//   auto dispatch(eve::tag_of<kyosu::asech> const&, C const& z) noexcept
+//   {
+//     return acosh(rec(z));
+//   }
 
-  template<typename C>
-  KYOSU_FORCEINLINE constexpr
-  auto dispatch(eve::tag_of<kyosu::acsch> const&, C const& z) noexcept
-  {
-    return sinh(rec(z));
-  }
+//   template<typename C>
+//   KYOSU_FORCEINLINE constexpr
+//   auto dispatch(eve::tag_of<kyosu::acsch> const&, C const& z) noexcept
+//   {
+//     return sinh(rec(z));
+//   }
 
-  template<typename C>
-  KYOSU_FORCEINLINE constexpr
-  auto dispatch(eve::tag_of<kyosu::asin> const&, C const& z) noexcept
-  {
-    auto s = sign(pure(z));
-    return -s*asinh(z*s);
-  }
+//   template<typename C>
+//   KYOSU_FORCEINLINE constexpr
+//   auto dispatch(eve::tag_of<kyosu::asin> const&, C const& z) noexcept
+//   {
+//     auto s = sign(pure(z));
+//     return -s*asinh(z*s);
+//   }
 
-  template<typename C>
-  KYOSU_FORCEINLINE constexpr
-  auto dispatch(eve::tag_of<kyosu::acos> const&, C const& z) noexcept
-  {
-    return sign(pure(z))*acosh(z);
-  }
+//   template<typename C>
+//   KYOSU_FORCEINLINE constexpr
+//   auto dispatch(eve::tag_of<kyosu::acos> const&, C const& z) noexcept
+//   {
+//     return sign(pure(z))*acosh(z);
+//   }
 
-  template<typename C>
-  KYOSU_FORCEINLINE constexpr
-  auto dispatch(eve::tag_of<kyosu::atan> const&, C const& z) noexcept
-  {
-    auto s = sign(pure(z));
-    return -s*atanh(z*s);
-  }
+//   template<typename C>
+//   KYOSU_FORCEINLINE constexpr
+//   auto dispatch(eve::tag_of<kyosu::atan> const&, C const& z) noexcept
+//   {
+//     auto s = sign(pure(z));
+//     return -s*atanh(z*s);
+//   }
 
-  template<typename C>
-  KYOSU_FORCEINLINE constexpr
-  auto dispatch(eve::tag_of<kyosu::acot> const&, C const& z) noexcept
-  {
-    auto s = sign(pure(z));
-    return s*acoth(z*s);
-  }
+//   template<typename C>
+//   KYOSU_FORCEINLINE constexpr
+//   auto dispatch(eve::tag_of<kyosu::acot> const&, C const& z) noexcept
+//   {
+//     auto s = sign(pure(z));
+//     return s*acoth(z*s);
+//   }
 
-  template<typename C>
-  KYOSU_FORCEINLINE constexpr
-  auto dispatch(eve::tag_of<kyosu::asec> const&, C const& z) noexcept
-  {
-    return acos(rec(z));
-  }
+//   template<typename C>
+//   KYOSU_FORCEINLINE constexpr
+//   auto dispatch(eve::tag_of<kyosu::asec> const&, C const& z) noexcept
+//   {
+//     return acos(rec(z));
+//   }
 
-  template<typename C>
-  KYOSU_FORCEINLINE constexpr
-  auto dispatch(eve::tag_of<kyosu::acsc> const&, C const& z) noexcept
-  {
-    return asin(rec(z));
-  }
+//   template<typename C>
+//   KYOSU_FORCEINLINE constexpr
+//   auto dispatch(eve::tag_of<kyosu::acsc> const&, C const& z) noexcept
+//   {
+//     return asin(rec(z));
+//   }
 
 }
